legacy/airfogsim/requirements.txt
absl-py==2.1.0
alabaster==1.0.0
attrs==24.2.0
Babel==2.15.0
certifi==2024.7.4
charset-normalizer==3.3.2
click==8.1.7
click-plugins==1.1.1
cligj==0.7.2
contourpy==1.3.0
cupy-cuda11x==13.3.0
cycler==0.12.1
docutils==0.21.2
fastrlock==0.8.2
filelock==3.15.4
fiona==1.9.6
fonttools==4.53.1
fsspec==2024.6.1
geopandas==0.14.4
grpcio==1.68.1
idna==3.7
imagesize==1.4.1
Jinja2==3.1.4
kiwisolver==1.4.5
Mako==1.3.5
Markdown==3.6
MarkupSafe==2.1.5
matplotlib==3.9.2
mpmath==1.3.0
networkx==3.3
numpy==1.24.4
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.20.5
nvidia-nvjitlink-cu12==12.5.82
nvidia-nvtx-cu12==12.1.105
osmnx==1.9.4
packaging==24.1
pandas==2.2.2
pdoc3==0.11.1
pillow==10.4.0
protobuf==5.29.2
Pygments==2.18.0
pyinstrument==5.0.0
pyparsing==3.1.4
pyproj==3.6.1
python-dateutil==2.9.0.post0
pytz==2024.1
PyYAML==6.0.1
requests==2.32.3
shapely==2.0.6
six==1.16.0
snowballstemmer==2.2.0
Sphinx==8.0.2
sphinxcontrib-applehelp==2.0.0
sphinxcontrib-devhelp==2.0.0
sphinxcontrib-htmlhelp==2.1.0
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==2.0.0
sphinxcontrib-serializinghtml==2.0.0
sumolib==1.20.0
sympy==1.13.1
tensorboard==2.18.0
tensorboard-data-server==0.7.2
tomli==2.0.1
torch==2.4.0
torchaudio==2.4.0
torchvision==0.19.0
tqdm==4.66.5
traci==1.20.0
triton==3.0.0
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.2
Werkzeug==3.1.3
